[open][neva_upstreamable][build] Fix unsigned vs signed comparison warning
:Release Notes:
Fix unsigned vs signed comparison warning

:Detailed Notes:
webruntime-1.0-r2 do_compile: [8/14] CXX obj/webos/weboswebruntime/webos_network_delegate.o
../../git/src/webos/browser/net/webos_network_delegate.cc: In member function 'void webos::WebOSNetworkDelegate::ParsePathsFromSettings(std::vector<std::__cxx11::basic_string<char> >&, std::istringstream&)':
../../git/src/webos/browser/net/webos_network_delegate.cc:96:18: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (path_pos != std::string::npos) {
         ~~~~~~~~~^~~~~~
../../git/src/webos/browser/net/webos_network_delegate.cc: In member function 'bool webos::WebOSNetworkDelegate::IsAccessAllowed(const net::URLRequest&, const base::FilePath&) const':
../../git/src/webos/browser/net/webos_network_delegate.cc:144:21: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < allowed_target_paths_.size(); ++i) {
                   ~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../../git/src/webos/browser/net/webos_network_delegate.cc:176:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < allowed_trusted_target_paths_.size(); ++i) {
                     ~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
:Testing Performed:
It builds well without warning.

:QA Notes:
N/A

:Issues Addressed:
N/A

Originally-Reviewed-on: http://gpro.lgsvl.com/218098
Reviewed-on: http://gpro.lgsvl.com/218542
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>

+++ src/webos/browser/net/webos_network_delegate.cc
@@ -96,7 +96,7 @@ void WebOSNetworkDelegate::ParsePathsFromSettings(
     std::getline(stream, str);
     if (str.find("size=") != std::string::npos)
       break;
-    int path_pos = str.find("=");
+    size_t path_pos = str.find("=");
     if (path_pos != std::string::npos) {
       paths.push_back(str.substr(path_pos + 1));
     }
@@ -117,7 +117,7 @@ bool WebOSNetworkDelegate::IsAccessAllowed(
   const base::FilePath stripped_path(path.StripTrailingSeparators());
 
   // 1. Resources in globally permitted paths
-  for (int i = 0; i < allowed_target_paths_.size(); ++i) {
+  for (size_t i = 0; i < allowed_target_paths_.size(); ++i) {
     const base::FilePath white_listed_path(allowed_target_paths_.at(i));
     // base::FilePath::operator== should probably handle trailing separators.
     if (white_listed_path == stripped_path || white_listed_path.IsParent(path))
@@ -149,7 +149,7 @@ bool WebOSNetworkDelegate::IsAccessAllowed(
 
   // 3. Resources for trusted app
   if (trust_level == "trusted") {
-    for (int i = 0; i < allowed_trusted_target_paths_.size(); ++i) {
+    for (size_t i = 0; i < allowed_trusted_target_paths_.size(); ++i) {
       base::FilePath config_path(allowed_trusted_target_paths_.at(i));
       // Strip trailingseparators, this allows using both /foo/ and /foo in
       // security_config file
